refactor: migrate memory_files to rara-memory crate
